treeval only takes expected value of highest value action

+++ include/helpers.h
@@ -2,4 +2,4 @@
 
 Node* generate();
 
-int treeval(Node *root);
+float treeval(Node *root);
